crud libro

+++ Libreria/CEntidades/Entidades/Libro.cs
@@ -8,7 +8,7 @@
         public string Titulo { get; set; }
         public DateTime FechaPublicacion { get; set; }
         public string Descripcion { get; set; }
-        public int PrecioVenta { get; set; }
+        public float PrecioVenta { get; set; }
 
         public ICollection<Autor> Autores { get; set; }
         public ICollection<Genero> Generos { get; set; }
